broke out validation into a React hook for greater organization
