+++ src/components/utils/private-route.jsx
@@ -1,10 +1,8 @@
-import React, { useState, useEffect } from 'react';
-import { useHistory, Route } from 'react-router-dom';
-import { axiosWithAuth } from '../../utilities/axiosWithAuth';
-import ls from 'local-storage';
-import axios from 'axios';
+import React, { useState, useEffect } from "react";
+import { useHistory, Route } from "react-router-dom";
+import useValidate from "../../hooks/useValidate";
 
-const qs = require('querystring');
+const qs = require("querystring");
 
 //const PrivateRoute = ({ component: Component, ...rest }) => {
 //   return (
@@ -22,55 +20,33 @@ const qs = require('querystring');
 // };
 
 function PrivateRoute({ component: Component, path, ...props }) {
-  const [response, setResponse] = useState(null);
-  const [requestMade, setRequestMade] = useState(false);
   const { push } = useHistory();
 
-  useEffect(() => {
-    const validate = async () => {
-      let res;
-      try {
-        if (ls.get('access_token')) {
-          if (!process.env.REACT_APP_FRONT_END_BASE_URL.includes('localhost')) {
-            const body = {
-              token: ls.get('access_token'),
-              token_type_hint: 'access_token',
-              client_id: process.env.REACT_APP_CLIENT_ID,
-            };
-            res = await axios.post(
-              'https://dev-36383529.okta.com/oauth2/default/v1/introspect',
-              qs.stringify(body)
-            );
-            setResponse(res);
-          } else {
-            res = await axiosWithAuth().post(
-              'http://localhost:5100/authenticate',
-              {}
-            );
-            setResponse(res);
-          }
-          setRequestMade(true);
-        }
-      } catch (err) {
-        console.dir(err);
-      }
-    };
+  let env = process.env.REACT_APP_FRONT_END_BASE_URL.includes("localhost")
+    ? "development"
+    : "production";
 
-    validate();
+  const { validate, response, requestMade } = useValidate();
+
+  useEffect(() => {
+    validate(env);
     // eslint-disable-next-line
   }, [path, Component]);
 
   if (!response) {
     return <p>...loading</p>;
   }
-  if (response.status.toString().match(/2[0-9][0-9]/) && response.data.active) {
+
+  if (
+    env === "production" &&
+    response.status.toString().match(/2[0-9][0-9]/) &&
+    response.data.active
+  ) {
     return <Route {...props} path={path} component={Component} />;
-  }
-  if (response.data.success) {
+  } else if (env == "development" && response.data.success) {
     return <Route {...props} path={path} component={Component} />;
-  }
-  if (!response.data.success && requestMade) {
-    push('/');
+  } else if (requestMade) {
+    push("/");
   }
   return null;
 }
